feat: add navigation for onboarding
